Update to keymap definition and helpers
Added extra bluetooth power mode to try, have had issues with connectivity.

Updated .keymap with helper keys, the definition is close enough to hillside definition so using it as base and expanding it.

Added the work on definions for the spanish layout made with https://github.com/joelspadin/zmk-locale-generator a great tool and will help a lot when planing keymaps